extract TokenCell
