+++ src/screens/Borrow/cells.tsx
@@ -4,26 +4,13 @@ import { PERCENT_DIGITS, TOKEN_FORMAT, USD_FORMAT } from "../../store/constants"
 import { IAsset } from "../../interfaces/account";
 import { IAssetDetailed } from "../../interfaces/asset";
 import { toUsd } from "../../store";
-import TokenIcon from "../../components/TokenIcon";
+import { TokenCell } from "../../components/Table/common/cells";
 
 interface CellProps {
 	rowData: IAsset & IAssetDetailed;
 }
 
-export const TokenCell = ({ rowData }) => {
-	const { symbol, price } = rowData;
-	return (
-		<Box display="flex">
-			<Box>
-				<TokenIcon icon={rowData?.icon} />
-			</Box>
-			<Box px="1rem">
-				<Box>{symbol}</Box>
-				<Box>{price ? `${rowData.price.usd.toLocaleString(undefined, USD_FORMAT)}` : "$-.-"}</Box>
-			</Box>
-		</Box>
-	);
-};
+export { TokenCell };
 
 export const APYCell = ({ rowData }: CellProps) => {
 	return <Box>{Number(rowData.borrow_apr).toFixed(PERCENT_DIGITS)}%</Box>;
